fix(test): dedupe test utils
